FX Correlation Analysis Application/requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0

# Data processing
pandas==2.2.0
numpy==1.26.4
pyarrow==15.0.0

# MT5 integration
MetaTrader5==5.0.45

# HTTP client (Binance)
httpx==0.27.0

# Configuration
python-dotenv==1.0.1

# File locking for cache safety
filelock==3.13.1

# Pydantic (already a FastAPI dep, pinned for clarity)
pydantic==2.6.0
